+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+* PyLucene >=9.12 required
 
 ## [3.2](https://pypi.org/project/lupyne/3.2/) - 2024-07-07
 ### Changed